requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
base58==2.1.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.2
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.7
loguru==0.7.2
multidict==6.0.5
py-near @ git+https://github.com/PowderDev/py-near@985d7e675d191dbc75abde71e311dbb199d4ced6
py_near_primitives==0.2.3
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
setuptools==69.5.1
sniffio==1.3.1
starlette==0.37.2
transaction==4.0
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4
zope.interface==6.3
